refactor(types): split identity.ts into entities/{member,member-photo,system}.ts
